Launch browser on darwin, linux, win32. Fail silently on the rest.
