--- goconvey_notdarwin.go
@@ -1,14 +1,0 @@
-// +build !darwin
-
-package main
-
-import (
-	"log"
-)
-
-func launchBrowser(host string, port int) {
-	log.Printf("Launching browser on %s:%d", host, port)
-	log.Printf("|--=-=-=--=-=-=-=--=--=--|")
-	log.Printf("|--NOT Implemented yet!--|")
-	log.Printf("|--=-=-=--=-=-=-==--==--=|")
-}
